MLMambaOut/requirements.txt
einops==0.8.0
matplotlib==3.8.4
nibabel==5.2.1
numpy==2.1.0
pandas==2.2.2
Pillow==10.4.0
scikit_learn==1.4.1.post1
timm==1.0.8
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.4
vision_mamba==0.1.0
zeta==0.7b1
